+++ calibration.py
@@ -1,18 +1,22 @@
-from StereoCalibration import StereoCalibration
+import glob
+import cv2 as cv
+import numpy as np
+import matplotlib.pyplot as plt
+
+from modules.StereoCalibration import StereoCalibration
+from modules.util import *
 
 # ================ PARAMETRES ========================
 
-patternSize=(9,6)
-squaresize=3.64e-2
+patternSize=(10,8)
+squaresize=2e-2
 # single_path='stereo/'
-stereo_path='stereo/'
-# single_detected_path='output/singles_detected/'
-stereo_detected_path='output/stereo_detected/'
-
+stereo_path='captures_flip/captures_calibration/'
+single_detected_path='output_flip/singles_detected/'
+stereo_detected_path='output_flip/stereo_detected/'
 # ====================================================
 
 obj = StereoCalibration(patternSize, squaresize)
-
-obj.calibrateStereo(stereo_path, stereo_detected_path,fisheye=False, calib_2_sets=False)
+obj.calibrateStereo(stereo_path, stereo_detected_path, single_detected_path, fisheye=False, calib_2_sets=False)
 obj.saveResultsXML()
-obj.reprojection('output/reprojection/')
+# obj.reprojection('output/reprojection/')
